backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
flask-pymongo==2.3.0
python-dotenv==1.0.0
